Implement "basic" Mastic
Add a new module to `poc` that implements the Mastic VDAF. This VDAF is
similar to Poplar1 except that instead of aggregating counters, Mastic
can aggregate any data type that is compatible with FlpGeneric.

Mastic operates in one of two "modes", depending on how the VIDPF is
configured. In "incremental mode", the FLP is checked when evaluating at
the first level of the VIDPF tree; in non-incremental mode, the FLP is
checked when evaluating the last level of the tree (all intermediate
levels are disabled).

When verifying the FLP, the encoded measurement (i.e., `beta`) is
computed by evaluating the VIDPF at both child nodes of the root and
summing the output shares. By one-hot verifiability, this sum is equal
to value claimed by the Client; by path verifiability, checking this
value is sufficient each node on the `alpha` path is guaranteed to have
the same value.

The code is incomplete but is sufficient to illustrate the basic
features we want. Some things to do:

* Add support for circuits with joint randomness

* Specify `PublicShare` and `InputShare`

* Decide whether to bind the query randomness to the VIDPF mode

* Implement message serialization