Pass measurement count to `Flp.decode()`
At the moment we simply pass `meas_count == 0` to `Flp.decode()` because
we don't know the actual value of `meas_count`. This makes `Mastic`
incompatible with any FLP that requires the measurement count at
decoding time.

To address this, plumb a counter for each measurement from VIDPF
evaluation to FLP unsharding. The value of the counter is `1` if the
prefix is on the `alpha`-path and `0` otherwise.

We accomplish this as follows. At generation time, prepend a field element to
each correction word that represents the counter. At evaluation time, we
compute our share of the counter and check that it is equal to `1` just
as PLASMA does.

Along the way, fix a bug in `Mastic.aggregate()` and `Mastic.unshard()`
whereby the accumulator vector was initialized to the wrong length.