Upgrade reference to base draft
In draft-08, SHAKE was replaced with TurboSHAKE, so upgrade the
dependency accordingly.

The latest version of the main branch has been checked out rather than
the tagged version (draft-irtf-cfrg-vdaf-08). This is because the latest
version resolves a technical issue regarding the implementation of
TurboSHAKE. This was resolved by upgrading PyCryptodomex to the latest
version:

$ sage -pip install --upgrade pycryptodomex

Version 3.20.0 is required.