Move eval proof computation to Mastic
Currently VIDPF computes the evaluation proof, which is largely
Mastic-specific. For instance, we would like to bind Mastic preparation
to the VDAF algorithm ID, as this allows for domain separation across
different weight types. But other applications of VIDPF may have
different requirements for domain separation.

This commit moves the eval proof computation to Mastic. VIDPF now just
computes the prefix tree and returns the root and leaves it up to
applications to do something with the result.

To maintain the current test coverage, we also make the following
changes:

1. Add a method on VIDPF, `test_eval()` that computes the hash of the
   node proofs and returns that instead of the prefix tree.

2. Move the `test_malformed_correction_word_payload()` to the Mastic
   tests. The node proof is not actually sensitive to changes to the
   payload correction word, but payload check is.