poc: Simplify the onehot check
For each node in the prefix tree that we traverse, we compute a "node
proof" according to [CP22]. We then combine the node proofs into the
"onehot proof" according to the following scheme.

Each node proof updates the onehot proof. An update resembles a step of
Merkle-Damgard compression. The main difference is that we XOR each
block (i.e., corrected node proof) with the previous hash (or IV) rather
than compress:

```
            node proof
               |
               v
current     +-----+     +------+     +-----+    updated
proof --+-->| XOR |---->| Hash |---->| XOR |--> proof
        |   +-----+     +------+     +-----+
        |                               ^
        |                               |
        +-------------------------------+
```

This construction is also from [CP22]. According to the authors, any
method of hashing the node proofs that is "secure against permutations
attacks" should suffice. All we really want is to confirm that the node
proofs computed by both servers are the same.

This commit replaces this bespoke construction with a single XOF
invocation, using the node proofs in breadth first order as the binder.

This change results in fewer XOF usage domain separators, so renumber
and reorder these separators as well.