poc: Let `beta` be the counter plus the encoded, truncated weight
Depending on the FLP, truncation may significantly reduce the length of
the measurement. For example, for the `Sum` type, the length of the
encoded measurement depends on the validity range, but the truncated
measurement is a single field element. We can therefore significantly
reduce the size of the correction words by programming the VIDPF with
the truncated weight instead of the full encoded weight.

To make this work, we secret share the encoded weight precisely as in
Prio3. We then extend the payload check by setting the root payload to
the counter plus the encoded truncated weight carried by the encoded
input share.

This change only saves communication cost when `mastic.flp.MEAS_LEN >
mastic.flp.OUTPUT_LEN` and `mastic.vidpf.BITS` is sufficiently large. Of
the currently defined variants, only `MasticSum` will see savings. For
`MasticCount`, `MasticSumVec`, and `MasticHistogram` the Leader's share
increases by `mastic.flp.MEAS_LEN` field elements.

Another benefit of this change is that the payload check detects
malformed correction word payloads at the first level of the tree.

Related clean up:

1. Add another usage domain separator for the helper's weight share and
   renumber.

2. Simplify the `MasticInputShare` type by defining classes for the
   leader and helper shares.

3. vidpf: Remove `get_beta_share()` as it's no longer used.