poc: Truncate the weight at each level but the first
First, extend VIDPF by adding an optional linear transform to the
payload. The transform may depend on the level of the tree we're
evaluating. Note that the length of each correction word depends on the
length of the transformed payload.

Second, modify Mastic by applying the following transformation: for all
but the first level of the tree, truncate the weight according to the
FLP. We skip the first level so that we can still do the range check at
the first level, but for each subsequent level, we don't need the full
weight share.

Depending on the FLP, this may significantly reduce the size of the
public share.